moved file locations to match
