+++ src/main/java/Player/Player.java
@@ -1,16 +1,18 @@
-package BlackJack.ComputerPlayer;
+package Player;
 
 import Card.Card;
 
+import java.util.ArrayList;
 import java.util.List;
 
-public class ComputerPlayer {
+public class Player {
+
     private String name;
     private int wallet = 500;
     private List<Card> currentHand;
 
-    public ComputerPlayer(List<Card> currentHandHand) {
-        this.currentHand = currentHandHand;
+    public Player(String name) {
+        this.name = name;
     }
 
     public String getName() {
@@ -41,7 +43,14 @@ public class ComputerPlayer {
         return totalValue;
     }
 
-    public void setCurrentHand(List<Card> currentHand) {
-        this.currentHand = currentHand;
+    public void setCurrentHand(List<Card> cards) {
+        this.currentHand = cards;
+    }
+
+    public void addCardsToHand(List<Card> cards) {
+        if (this.currentHand == null) {
+            this.currentHand = new ArrayList<>();
+        }
+        this.currentHand.addAll(cards);
     }
 }
